feat: implement job position management with integrated salary structure configuration

+++ worksphere/src/main/java/com/ucocs/worksphere/service/JobPositionService.java
@@ -2,7 +2,10 @@ package com.ucocs.worksphere.service;
 
 import com.ucocs.worksphere.entity.JobPosition;
 import com.ucocs.worksphere.exception.ResourceNotFoundException;
+import com.ucocs.worksphere.repository.EmployeeRepository;
 import com.ucocs.worksphere.repository.JobPositionRepository;
+import com.ucocs.worksphere.repository.SalaryStructureRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,9 +14,16 @@ import java.util.UUID;
 @Service
 public class JobPositionService {
     private final JobPositionRepository jobPositionRepository;
+    private final EmployeeRepository employeeRepository;
+    private final SalaryStructureRepository salaryStructureRepository;
 
-    public JobPositionService(JobPositionRepository jobPositionRepository) {
+    public JobPositionService(JobPositionRepository jobPositionRepository,
+                              EmployeeRepository employeeRepository,
+                              SalaryStructureRepository salaryStructureRepository) {
         this.jobPositionRepository = jobPositionRepository;
+        this.employeeRepository = employeeRepository;
+        this.salaryStructureRepository = salaryStructureRepository;
+
     }
 
     public JobPosition createJobPosition(JobPosition jobPosition) {
@@ -32,9 +42,15 @@ public class JobPositionService {
         return jobPositionRepository.save(jobPosition);
     }
 
+    @Transactional
     public void deleteJobPosition(UUID id) {
         JobPosition jobPosition = jobPositionRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Job Position not found"));
+        if(employeeRepository.existsByJobPositionId(id))
+        {
+            throw new ResourceNotFoundException("Employee exist with this job title");
+        }
+        salaryStructureRepository.deleteByJobPositionId(id);
         jobPositionRepository.delete(jobPosition);
     }
